Add code formatting via prettier
Prettier is an opinionated JS code formatter.

There are several ways to integrate it into your workflow.

One way is to use eslint-plugin-prettier, which does several things:
- disables eslint's style checking rules
- tells eslint to run prettier
- reports style violations as eslint errors that can be fixed with --fix

The advantage to this approach is that all checking is done via a single command, and all errors (formatting or otherwise) are reported in the linter output.  You can enable your editor's eslint "fix on save" option to automatically format code on save.

We elect not to use this option, as we prefer standard JS style, which disagrees with prettier on one or two rules.  Instead, we leverage prettier-eslint, which runs prettier and then passes its output to eslint --fix.  When eslint is configured to use standard, this re-applies the standard formatting rules that prettier doesn't adopt.

In order for this to work, we have to abandon the standard linter in favor of using eslint directly with a standard JS configuration (see the previous commit).